refactor(web): split daemon web modules
Split daemon Web code into dedicated modules for install, serve, dashboard, and chat while keeping the public module name as web_cli.

Also wire the browser-facing Web icon through web/index.html.
