+++ crates/daemon/src/web/auth.rs
