+++ crates/daemon/src/web/debug_console.rs
