+++ crates/daemon/src/web/mod.rs
@@ -40,15 +40,31 @@ use tokio::{
 use crate::{CliResult, mvp, with_graceful_shutdown};
 
 mod auth;
+mod chat;
+mod dashboard;
 mod debug_console;
+mod install;
 mod onboarding;
+mod serve;
 
 use auth::{
     build_clear_pairing_cookie, build_clear_same_origin_session_cookie, build_pairing_cookie,
     build_same_origin_session_cookie, extract_allowed_local_origin, extract_request_token,
     request_is_authenticated, require_local_token, require_same_origin_write_origin,
 };
+use chat::{
+    chat_history, chat_sessions, chat_turn, chat_turn_stream, create_chat_session,
+    delete_chat_session,
+};
+use dashboard::{
+    dashboard_config, dashboard_connectivity, dashboard_providers, dashboard_runtime,
+    dashboard_summary, dashboard_tools,
+};
 use debug_console::{dashboard_debug_console, record_debug_operation};
+use install::{
+    default_web_install_dir, run_web_install, run_web_remove, run_web_status, web_install_dist_dir,
+};
+use serve::run_web_serve;
 
 #[derive(Subcommand, Debug)]
 pub enum WebCommand {
@@ -450,548 +466,6 @@ pub async fn run_web_command(command: WebCommand) -> CliResult<()> {
         WebCommand::Status => run_web_status(),
         WebCommand::Remove { force } => run_web_remove(force),
     }
-}
-
-async fn run_web_serve(
-    config_path: Option<&str>,
-    bind: &str,
-    static_root: Option<&str>,
-) -> CliResult<()> {
-    let (local_token, local_token_path) = resolve_local_web_token()
-        .map_err(|error| format!("initialize local web api token failed: {}", error.message))?;
-    let token_path_display = local_token_path.display().to_string();
-    let explicit_static_root = resolve_static_root(static_root)?;
-    let resolved_static_root = if explicit_static_root.is_some() {
-        explicit_static_root
-    } else {
-        let auto_dist = web_install_dist_dir(&default_web_install_dir());
-        if auto_dist.join("index.html").is_file() {
-            Some(auto_dist)
-        } else {
-            None
-        }
-    };
-    let web_install_mode = if resolved_static_root.is_some() {
-        "same_origin_static"
-    } else {
-        "api_only"
-    };
-    let address: SocketAddr = bind
-        .parse()
-        .map_err(|error| format!("invalid web bind address `{bind}`: {error}"))?;
-    if web_install_mode == "same_origin_static" && !address.ip().is_loopback() {
-        return Err(format!(
-            "same-origin static mode only supports loopback binds, got `{bind}`"
-        ));
-    }
-    let exact_origin = matches!(
-        address.ip(),
-        std::net::IpAddr::V4(_) | std::net::IpAddr::V6(_)
-    )
-    .then(|| format!("http://{address}"));
-    let state = Arc::new(WebApiState {
-        config_path: config_path.map(str::to_owned),
-        local_token,
-        local_token_path,
-        web_install_mode,
-        exact_origin,
-        static_root: resolved_static_root.clone(),
-        turn_streams: Mutex::new(HashMap::new()),
-        debug_state: StdMutex::new(DebugConsoleRuntimeState::default()),
-    });
-    let public_api = Router::new()
-        .route("/meta", get(meta))
-        .route("/onboard/status", get(onboarding::onboard_status))
-        .route(
-            "/onboard/pairing/auto",
-            post(onboarding::onboard_pairing_auto),
-        )
-        .route(
-            "/onboard/pairing/clear",
-            post(onboarding::onboard_pairing_clear),
-        )
-        .with_state(state.clone());
-    let protected_api = Router::new()
-        .route("/onboard/provider", post(onboarding::onboard_provider))
-        .route(
-            "/onboard/provider/apply",
-            post(onboarding::onboard_provider_apply),
-        )
-        .route(
-            "/onboard/preferences",
-            post(onboarding::onboard_preferences),
-        )
-        .route("/onboard/validate", post(onboarding::onboard_validate))
-        .route("/dashboard/summary", get(dashboard_summary))
-        .route("/dashboard/providers", get(dashboard_providers))
-        .route("/dashboard/runtime", get(dashboard_runtime))
-        .route("/dashboard/connectivity", get(dashboard_connectivity))
-        .route("/dashboard/config", get(dashboard_config))
-        .route("/dashboard/tools", get(dashboard_tools))
-        .route("/dashboard/debug-console", get(dashboard_debug_console))
-        .route(
-            "/chat/sessions",
-            get(chat_sessions).post(create_chat_session),
-        )
-        .route("/chat/sessions/{id}", delete(delete_chat_session))
-        .route("/chat/sessions/{id}/turn", post(chat_turn))
-        .route(
-            "/chat/sessions/{id}/turns/{turn_id}/stream",
-            get(chat_turn_stream),
-        )
-        .route("/chat/sessions/{id}/history", get(chat_history))
-        .layer(middleware::from_fn_with_state(
-            state.clone(),
-            require_same_origin_write_origin,
-        ))
-        .layer(middleware::from_fn_with_state(
-            state.clone(),
-            require_local_token,
-        ))
-        .with_state(state.clone());
-    let app = Router::new()
-        .route("/healthz", get(healthz))
-        .nest("/api", public_api.merge(protected_api))
-        .fallback(get(serve_web_static))
-        .layer(middleware::from_fn(local_web_cors))
-        .with_state(state);
-
-    let listener = tokio::net::TcpListener::bind(address)
-        .await
-        .map_err(|error| format!("bind web api on {bind} failed: {error}"))?;
-
-    println!("loongclaw web api listening on http://{address}");
-    println!("loongclaw web api local token path: {token_path_display}");
-    if let Some(static_root) = resolved_static_root.as_ref() {
-        println!(
-            "loongclaw web ui same-origin static root: {}",
-            static_root.display()
-        );
-    }
-    with_graceful_shutdown(async move {
-        axum::serve(listener, app)
-            .await
-            .map_err(|error| format!("web api serve failed: {error}"))
-    })
-    .await
-}
-
-async fn healthz() -> Json<ApiEnvelope<HealthPayload>> {
-    Json(ApiEnvelope {
-        ok: true,
-        data: HealthPayload { status: "ok" },
-    })
-}
-
-async fn local_web_cors(request: Request, next: Next) -> Response {
-    let allowed_origin = extract_allowed_local_origin(request.headers());
-    if request.method() == Method::OPTIONS {
-        return with_cors_headers(
-            StatusCode::NO_CONTENT.into_response(),
-            allowed_origin.as_deref(),
-        );
-    }
-
-    let response = next.run(request).await;
-    with_cors_headers(response, allowed_origin.as_deref())
-}
-
-fn with_cors_headers(mut response: Response, allowed_origin: Option<&str>) -> Response {
-    if let Some(origin) = allowed_origin
-        && let Ok(value) = HeaderValue::from_str(origin)
-    {
-        response
-            .headers_mut()
-            .insert(ACCESS_CONTROL_ALLOW_ORIGIN, value);
-        response.headers_mut().insert(
-            ACCESS_CONTROL_ALLOW_CREDENTIALS,
-            HeaderValue::from_static("true"),
-        );
-        response
-            .headers_mut()
-            .insert(VARY, HeaderValue::from_static("Origin"));
-    }
-    response.headers_mut().insert(
-        ACCESS_CONTROL_ALLOW_METHODS,
-        HeaderValue::from_static("GET, POST, DELETE, OPTIONS"),
-    );
-    response.headers_mut().insert(
-        ACCESS_CONTROL_ALLOW_HEADERS,
-        HeaderValue::from_static("content-type, authorization, x-loongclaw-token"),
-    );
-    response
-}
-
-async fn meta(State(state): State<Arc<WebApiState>>) -> Json<ApiEnvelope<MetaPayload>> {
-    Json(ApiEnvelope {
-        ok: true,
-        data: MetaPayload {
-            app_version: env!("CARGO_PKG_VERSION").to_owned(),
-            api_version: "v1",
-            web_install_mode: state.web_install_mode,
-            supported_locales: ["en", "zh-CN"],
-            default_locale: "en",
-            auth: MetaAuthPayload {
-                required: true,
-                scheme: if state.web_install_mode == "same_origin_static" {
-                    "cookie"
-                } else {
-                    "bearer"
-                },
-                header: if state.web_install_mode == "same_origin_static" {
-                    "Cookie"
-                } else {
-                    "Authorization"
-                },
-                token_path: if state.web_install_mode == "same_origin_static" {
-                    String::new()
-                } else {
-                    state.local_token_path.display().to_string()
-                },
-                token_env: if state.web_install_mode == "same_origin_static" {
-                    ""
-                } else {
-                    WEB_API_TOKEN_ENV
-                },
-                mode: if state.web_install_mode == "same_origin_static" {
-                    "same_origin_session"
-                } else {
-                    "local_token"
-                },
-            },
-        },
-    })
-}
-
-async fn dashboard_summary(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<DashboardSummaryPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: DashboardSummaryPayload {
-            runtime_status: "ready",
-            active_provider: snapshot.config.active_provider_id().map(str::to_owned),
-            active_model: snapshot.config.provider.model.clone(),
-            memory_backend: "sqlite",
-            session_count: snapshot.sessions.len(),
-            web_install_mode: state.web_install_mode,
-        },
-    }))
-}
-
-async fn dashboard_providers(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<DashboardProvidersPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: DashboardProvidersPayload {
-            active_provider: snapshot.config.active_provider_id().map(str::to_owned),
-            items: build_provider_items(&snapshot.config),
-        },
-    }))
-}
-
-async fn dashboard_runtime(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<DashboardRuntimePayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: DashboardRuntimePayload {
-            status: "ready",
-            source: "local_daemon",
-            config_path: snapshot.resolved_path.display().to_string(),
-            memory_backend: snapshot.config.memory.resolved_backend().as_str(),
-            memory_mode: snapshot.config.memory.resolved_mode().as_str(),
-            ingest_mode: snapshot.config.memory.ingest_mode.as_str(),
-            web_install_mode: state.web_install_mode,
-            active_provider: snapshot.config.active_provider_id().map(str::to_owned),
-            active_model: snapshot.config.provider.model.clone(),
-            acp_enabled: snapshot.config.acp.enabled,
-            strict_memory: !snapshot.config.memory.effective_fail_open(),
-        },
-    }))
-}
-
-async fn dashboard_connectivity(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<DashboardConnectivityPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    let endpoint = snapshot.config.provider.endpoint();
-    let parsed = reqwest::Url::parse(&endpoint).map_err(|error| {
-        WebApiError::internal(format!(
-            "parse provider endpoint for connectivity failed: {error}"
-        ))
-    })?;
-    let host = parsed
-        .host_str()
-        .ok_or_else(|| WebApiError::internal("provider endpoint host was missing"))?
-        .to_owned();
-    let port = parsed.port_or_known_default().unwrap_or(443);
-    let dns_addresses = resolve_provider_host_addresses(host.as_str(), port).await;
-    let fake_ip_detected = dns_addresses
-        .iter()
-        .any(|address| is_fake_ip_address(address));
-    let proxy_env_detected = has_proxy_environment();
-    let (probe_status, probe_status_code) = probe_provider_endpoint(endpoint.as_str()).await;
-    let degraded = fake_ip_detected || probe_status != "reachable";
-    let recommendation = if fake_ip_detected {
-        Some("direct_host_and_fake_ip_filter")
-    } else if probe_status != "reachable" {
-        Some("check_network_route")
-    } else {
-        None
-    };
-
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: DashboardConnectivityPayload {
-            status: if degraded { "degraded" } else { "healthy" },
-            endpoint,
-            host,
-            dns_addresses,
-            probe_status,
-            probe_status_code,
-            fake_ip_detected,
-            proxy_env_detected,
-            recommendation,
-        },
-    }))
-}
-
-async fn dashboard_config(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<DashboardConfigPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    let active_provider = build_provider_items(&snapshot.config)
-        .into_iter()
-        .find(|item| item.enabled);
-
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: DashboardConfigPayload {
-            active_provider: snapshot.config.active_provider_id().map(str::to_owned),
-            last_provider: snapshot.config.last_provider.clone(),
-            model: snapshot.config.provider.model.clone(),
-            endpoint: snapshot.config.provider.endpoint(),
-            api_key_configured: active_provider
-                .as_ref()
-                .map(|item| item.api_key_configured)
-                .unwrap_or(false),
-            api_key_masked: active_provider.and_then(|item| item.api_key_masked),
-            personality: prompt_personality_id(snapshot.config.cli.resolved_personality())
-                .to_owned(),
-            prompt_mode: if snapshot.config.cli.uses_native_prompt_pack() {
-                "native_prompt_pack"
-            } else {
-                "inline_prompt"
-            },
-            prompt_addendum_configured: snapshot
-                .config
-                .cli
-                .system_prompt_addendum
-                .as_deref()
-                .map(str::trim)
-                .is_some_and(|value| !value.is_empty()),
-            memory_profile: snapshot
-                .config
-                .memory
-                .resolved_profile()
-                .as_str()
-                .to_owned(),
-            memory_system: snapshot.config.memory.resolved_system().as_str(),
-            sqlite_path: snapshot
-                .config
-                .memory
-                .resolved_sqlite_path()
-                .display()
-                .to_string(),
-            file_root: snapshot
-                .config
-                .tools
-                .resolved_file_root()
-                .display()
-                .to_string(),
-            sliding_window: snapshot.config.memory.sliding_window,
-            summary_max_chars: snapshot.config.memory.summary_max_chars,
-        },
-    }))
-}
-
-async fn dashboard_tools(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<DashboardToolsPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    let tool_runtime = mvp::tools::runtime_config::ToolRuntimeConfig::from_loongclaw_config(
-        &snapshot.config,
-        None,
-    );
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: DashboardToolsPayload {
-            approval_mode: approval_mode_label(snapshot.config.tools.approval.mode).to_owned(),
-            shell_default_mode: snapshot.config.tools.shell_default_mode.clone(),
-            shell_allow_count: snapshot.config.tools.shell_allow.len(),
-            shell_deny_count: snapshot.config.tools.shell_deny.len(),
-            items: build_tool_items(&snapshot.config, &tool_runtime),
-        },
-    }))
-}
-
-async fn chat_sessions(
-    State(state): State<Arc<WebApiState>>,
-) -> Result<Json<ApiEnvelope<ChatSessionsPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    let items = snapshot
-        .sessions
-        .iter()
-        .map(|session| ChatSessionItemPayload {
-            id: session.id.clone(),
-            title: session.title.clone(),
-            updated_at: format_timestamp(session.latest_turn_ts),
-        })
-        .collect();
-
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: ChatSessionsPayload { items },
-    }))
-}
-
-async fn create_chat_session(
-    Json(payload): Json<CreateChatSessionRequest>,
-) -> Json<ApiEnvelope<CreateChatSessionPayload>> {
-    let session_id = payload
-        .title
-        .as_deref()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(session_id_from_title)
-        .unwrap_or_else(generate_session_id);
-
-    Json(ApiEnvelope {
-        ok: true,
-        data: CreateChatSessionPayload { session_id },
-    })
-}
-
-async fn chat_history(
-    State(state): State<Arc<WebApiState>>,
-    Path(id): Path<String>,
-) -> Result<Json<ApiEnvelope<ChatHistoryPayload>>, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    let history = load_visible_session_messages(&snapshot.memory_config, &id, 128, 256)?;
-
-    if history.is_empty() {
-        return Err(WebApiError::not_found(format!(
-            "session `{id}` was not found in sqlite memory"
-        )));
-    }
-
-    let messages = history
-        .into_iter()
-        .enumerate()
-        .map(|(index, turn)| ChatMessagePayload {
-            id: format!("{id}:{index}"),
-            role: turn.role,
-            content: turn.content,
-            created_at: format_timestamp(turn.ts),
-        })
-        .collect();
-
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: ChatHistoryPayload {
-            session_id: id,
-            messages,
-        },
-    }))
-}
-
-async fn delete_chat_session(
-    State(state): State<Arc<WebApiState>>,
-    Path(id): Path<String>,
-) -> Result<StatusCode, WebApiError> {
-    let snapshot = load_web_snapshot(state.as_ref())?;
-    mvp::memory::clear_session_direct(&id, &snapshot.memory_config)
-        .map_err(WebApiError::internal)?;
-    Ok(StatusCode::NO_CONTENT)
-}
-
-async fn chat_turn(
-    State(state): State<Arc<WebApiState>>,
-    Path(id): Path<String>,
-    Json(payload): Json<ChatTurnRequest>,
-) -> Result<Json<ApiEnvelope<ChatTurnPayload>>, WebApiError> {
-    let input = payload.input.trim();
-    if input.is_empty() {
-        return Err(WebApiError {
-            status: StatusCode::BAD_REQUEST,
-            code: "invalid_request",
-            message: "chat turn input must not be empty".to_owned(),
-        });
-    }
-
-    let turn_id = generate_turn_id();
-    let (sender, receiver) = mpsc::unbounded_channel();
-
-    state
-        .turn_streams
-        .lock()
-        .await
-        .insert(turn_id.clone(), receiver);
-
-    let state_for_turn = state.clone();
-    let session_id = id.clone();
-    let turn_id_for_task = turn_id.clone();
-    let input_owned = input.to_owned();
-    tokio::spawn(async move {
-        let _ = run_chat_turn_stream(
-            state_for_turn,
-            session_id,
-            turn_id_for_task,
-            input_owned,
-            sender,
-        )
-        .await;
-    });
-
-    Ok(Json(ApiEnvelope {
-        ok: true,
-        data: ChatTurnPayload {
-            session_id: id,
-            turn_id,
-            status: "accepted",
-        },
-    }))
-}
-
-async fn chat_turn_stream(
-    State(state): State<Arc<WebApiState>>,
-    Path((_session_id, turn_id)): Path<(String, String)>,
-) -> Result<Response, WebApiError> {
-    let receiver = state
-        .turn_streams
-        .lock()
-        .await
-        .remove(&turn_id)
-        .ok_or_else(|| WebApiError::not_found(format!("turn `{turn_id}` was not found")))?;
-
-    let body_stream = stream::unfold(receiver, |mut receiver| async move {
-        receiver
-            .recv()
-            .await
-            .map(|line| (Ok::<String, Infallible>(format!("{line}\n")), receiver))
-    });
-
-    let mut response = Response::new(Body::from_stream(body_stream));
-    response.headers_mut().insert(
-        CONTENT_TYPE,
-        HeaderValue::from_static("application/x-ndjson"),
-    );
-    Ok(response)
 }
 
 struct WebSnapshot {
@@ -1408,165 +882,7 @@ fn resolve_local_web_token() -> Result<(String, PathBuf), WebApiError> {
 
 // ── Web install helpers ──────────────────────────────────────────────────────
 
-fn default_web_install_dir() -> PathBuf {
-    mvp::config::default_loongclaw_home().join("web")
-}
-
-fn web_install_dist_dir(install_dir: &FsPath) -> PathBuf {
-    install_dir.join("dist")
-}
-
-fn web_install_manifest_path(install_dir: &FsPath) -> PathBuf {
-    install_dir.join("install.json")
-}
-
-fn copy_dir_all(src: &FsPath, dst: &FsPath) -> CliResult<()> {
-    for entry in
-        fs::read_dir(src).map_err(|error| format!("failed to read `{}`: {error}", src.display()))?
-    {
-        let entry = entry.map_err(|error| format!("failed to read directory entry: {error}"))?;
-        let src_path = entry.path();
-        let dst_path = dst.join(entry.file_name());
-        if src_path.is_dir() {
-            fs::create_dir_all(&dst_path)
-                .map_err(|error| format!("failed to create `{}`: {error}", dst_path.display()))?;
-            copy_dir_all(&src_path, &dst_path)?;
-        } else {
-            fs::copy(&src_path, &dst_path).map_err(|error| {
-                format!(
-                    "failed to copy `{}` to `{}`: {error}",
-                    src_path.display(),
-                    dst_path.display()
-                )
-            })?;
-        }
-    }
-    Ok(())
-}
-
-fn run_web_install(source: &str) -> CliResult<()> {
-    let source_path = PathBuf::from(source);
-    if !source_path.exists() {
-        return Err(format!(
-            "source path `{}` does not exist",
-            source_path.display()
-        ));
-    }
-    if !source_path.is_dir() {
-        return Err(format!(
-            "source path `{}` is not a directory",
-            source_path.display()
-        ));
-    }
-    if !source_path.join("index.html").is_file() {
-        return Err(format!(
-            "source path `{}` is missing `index.html` — run `npm run build` first",
-            source_path.display()
-        ));
-    }
-
-    let install_dir = default_web_install_dir();
-    let dist_dir = web_install_dist_dir(&install_dir);
-
-    if dist_dir.exists() {
-        fs::remove_dir_all(&dist_dir).map_err(|error| {
-            format!(
-                "failed to remove existing install at `{}`: {error}",
-                dist_dir.display()
-            )
-        })?;
-    }
-    fs::create_dir_all(&dist_dir).map_err(|error| {
-        format!(
-            "failed to create install directory `{}`: {error}",
-            dist_dir.display()
-        )
-    })?;
-
-    copy_dir_all(&source_path, &dist_dir)?;
-
-    let canonical_source = source_path
-        .canonicalize()
-        .unwrap_or_else(|_| source_path.clone());
-    let manifest = WebInstallManifest {
-        installed_at: OffsetDateTime::now_utc()
-            .format(&Rfc3339)
-            .unwrap_or_default(),
-        source_path: canonical_source.display().to_string(),
-        install_dir: install_dir.display().to_string(),
-    };
-    let manifest_json = serde_json::to_string_pretty(&manifest)
-        .map_err(|error| format!("failed to serialize install manifest: {error}"))?;
-    fs::write(web_install_manifest_path(&install_dir), manifest_json)
-        .map_err(|error| format!("failed to write install manifest: {error}"))?;
-
-    println!("Web Console installed to: {}", dist_dir.display());
-    println!("Run `loongclaw web serve` to start the same-origin Web Console.");
-    Ok(())
-}
-
-fn run_web_status() -> CliResult<()> {
-    let install_dir = default_web_install_dir();
-    let manifest_path = web_install_manifest_path(&install_dir);
-    let dist_dir = web_install_dist_dir(&install_dir);
-
-    if !manifest_path.exists() {
-        println!("Web Console: not installed");
-        println!("Run `loongclaw web install --source <path/to/web/dist>` to install.");
-        return Ok(());
-    }
-
-    let manifest_raw = fs::read_to_string(&manifest_path)
-        .map_err(|error| format!("failed to read install manifest: {error}"))?;
-    let manifest: WebInstallManifest = serde_json::from_str(&manifest_raw)
-        .map_err(|error| format!("failed to parse install manifest: {error}"))?;
-
-    let assets_ok = dist_dir.join("index.html").is_file();
-    println!("Web Console: installed");
-    println!("Install dir:  {}", manifest.install_dir);
-    println!("Installed at: {}", manifest.installed_at);
-    println!("Source:       {}", manifest.source_path);
-    println!(
-        "Assets:       {}",
-        if assets_ok {
-            "ok"
-        } else {
-            "missing (dist/index.html not found — re-run `web install`)"
-        }
-    );
-    Ok(())
-}
-
-fn run_web_remove(force: bool) -> CliResult<()> {
-    let install_dir = default_web_install_dir();
-    let manifest_path = web_install_manifest_path(&install_dir);
-    let dist_dir = web_install_dist_dir(&install_dir);
-
-    if !manifest_path.exists() && !dist_dir.exists() {
-        println!("Web Console: not installed, nothing to remove.");
-        return Ok(());
-    }
-
-    if !force {
-        println!("This will remove: {}", install_dir.display());
-        println!("Re-run with --force to confirm removal.");
-        return Ok(());
-    }
-
-    if dist_dir.exists() {
-        fs::remove_dir_all(&dist_dir)
-            .map_err(|error| format!("failed to remove `{}`: {error}", dist_dir.display()))?;
-    }
-    if manifest_path.exists() {
-        fs::remove_file(&manifest_path)
-            .map_err(|error| format!("failed to remove `{}`: {error}", manifest_path.display()))?;
-    }
-
-    println!("Web Console removed from: {}", install_dir.display());
-    Ok(())
-}
-
-// ── Config / static root helpers ─────────────────────────────────────────────
+// ==== Config / static root helpers ====
 
 fn resolve_web_config_path(state: &WebApiState) -> PathBuf {
     state
@@ -1574,119 +890,6 @@ fn resolve_web_config_path(state: &WebApiState) -> PathBuf {
         .as_deref()
         .map(PathBuf::from)
         .unwrap_or_else(mvp::config::default_config_path)
-}
-
-fn resolve_static_root(static_root: Option<&str>) -> CliResult<Option<PathBuf>> {
-    let Some(raw_root) = static_root else {
-        return Ok(None);
-    };
-    let root = PathBuf::from(raw_root);
-    if !root.exists() {
-        return Err(format!(
-            "web static root `{}` does not exist",
-            root.display()
-        ));
-    }
-    if !root.is_dir() {
-        return Err(format!(
-            "web static root `{}` is not a directory",
-            root.display()
-        ));
-    }
-    let index_path = root.join("index.html");
-    if !index_path.is_file() {
-        return Err(format!(
-            "web static root `{}` is missing `index.html`",
-            root.display()
-        ));
-    }
-    Ok(Some(root))
-}
-
-async fn serve_web_static(
-    State(state): State<Arc<WebApiState>>,
-    uri: Uri,
-) -> Result<Response, WebApiError> {
-    let Some(static_root) = state.static_root.as_ref() else {
-        return Err(WebApiError::not_found("not found"));
-    };
-
-    let request_path = uri.path();
-    let candidate = match resolve_static_asset_path(static_root, request_path) {
-        Some(path) => path,
-        None => return Err(WebApiError::not_found("not found")),
-    };
-    let effective_path = if candidate.is_file() {
-        candidate
-    } else if is_asset_like_path(request_path) {
-        return Err(WebApiError::not_found("not found"));
-    } else {
-        static_root.join("index.html")
-    };
-    let bytes = fs::read(&effective_path).map_err(|error| {
-        WebApiError::internal(format!(
-            "read web static asset `{}` failed: {error}",
-            effective_path.display()
-        ))
-    })?;
-    let mut response = Response::builder()
-        .status(StatusCode::OK)
-        .header(CONTENT_TYPE, detect_static_content_type(&effective_path))
-        .body(Body::from(bytes))
-        .map_err(|error| WebApiError::internal(format!("build static response failed: {error}")))?;
-    if state.web_install_mode == "same_origin_static" {
-        response.headers_mut().append(
-            SET_COOKIE,
-            build_same_origin_session_cookie(state.local_token.as_str())?,
-        );
-    }
-    Ok(response)
-}
-
-fn resolve_static_asset_path(static_root: &FsPath, request_path: &str) -> Option<PathBuf> {
-    let trimmed = request_path.trim_start_matches('/');
-    if trimmed.is_empty() {
-        return Some(static_root.join("index.html"));
-    }
-
-    let relative = FsPath::new(trimmed);
-    let mut resolved = static_root.to_path_buf();
-    for component in relative.components() {
-        match component {
-            std::path::Component::Normal(segment) => resolved.push(segment),
-            std::path::Component::Prefix(_)
-            | std::path::Component::RootDir
-            | std::path::Component::CurDir
-            | std::path::Component::ParentDir => return None,
-        }
-    }
-    if resolved.is_dir() {
-        Some(resolved.join("index.html"))
-    } else {
-        Some(resolved)
-    }
-}
-
-fn is_asset_like_path(request_path: &str) -> bool {
-    FsPath::new(request_path.trim_start_matches('/'))
-        .extension()
-        .is_some()
-}
-
-fn detect_static_content_type(path: &FsPath) -> &'static str {
-    match path.extension().and_then(|value| value.to_str()) {
-        Some("html") => "text/html; charset=utf-8",
-        Some("js") => "application/javascript; charset=utf-8",
-        Some("css") => "text/css; charset=utf-8",
-        Some("svg") => "image/svg+xml",
-        Some("json") => "application/json; charset=utf-8",
-        Some("webmanifest") => "application/manifest+json; charset=utf-8",
-        Some("png") => "image/png",
-        Some("jpg") | Some("jpeg") => "image/jpeg",
-        Some("woff") => "font/woff",
-        Some("woff2") => "font/woff2",
-        _ => "application/octet-stream",
-    }
 }
 
 fn format_timestamp(unix_seconds: i64) -> String {
