+++ crates/daemon/src/web/onboarding.rs
