+++ web/index.html
@@ -1,8 +1,9 @@
-<!doctype html>
+﻿<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
+    <link rel="icon" type="image/svg+xml" href="/src/assets/brand/icon.svg" />
     <meta
       name="description"
       content="LoongClaw Web Console for chat and runtime visibility."
@@ -14,3 +15,4 @@
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
+
